@@ -1,9 +1,8 @@
 package com.zane.colorballch.base;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 /**
  * Created by shizhang on 2017/10/29.
  */
